fix(frontend): break circular imports in mcp-servers, assets, and Workflows
madge found 6 circular dependencies, each caused by the same pattern: a
component file defined a shared type, and its own child components imported
that type back from it. Cycles make module init order fragile (values can be
undefined at import time depending on entry point) and make each file harder
to reason about in isolation.

Moved the shared types into a sibling types.ts in each directory
(McpServer, Trigger, Asset/AssetVersion) and re-exported them from the
original files for backward compatibility with existing consumers.
