+++ app/frontend/pages/Projects/Workflows/TriggerFormPanel.tsx
@@ -6,7 +6,7 @@ import { useCallback, useState } from 'react';
 import { apiFetch } from 'shared/lib/apiFetch';
 import { apiV1ProjectWorkflowTriggerPath, apiV1ProjectWorkflowTriggersPath } from 'shared/routes';
 
-import type { Trigger } from './TriggersTab';
+import type { Trigger } from './types';
 
 interface ColumnOption {
   id: number;
