+++ app/frontend/pages/Projects/Workflows/TriggersTab.tsx
@@ -16,6 +16,9 @@ import { apiFetch } from 'shared/lib/apiFetch';
 import { apiV1ProjectWorkflowTriggerPath, apiV1ProjectWorkflowTriggersPath } from 'shared/routes';
 
 import { TriggerFormPanel } from './TriggerFormPanel';
+import type { Trigger } from './types';
+
+export type { Trigger } from './types';
 
 interface ColumnOption {
   id: number;
@@ -26,23 +29,6 @@ interface ColumnOption {
 interface StepOption {
   id: number;
   name: string;
-}
-
-export interface Trigger {
-  id: number;
-  kind: string;
-  event_type: string;
-  name?: string | null;
-  trigger_mode?: string;
-  cooldown_seconds?: number;
-  enabled?: boolean;
-  column_name?: string;
-  board_column_id?: number;
-  subject_policy?: string;
-  subject_column_id?: number | null;
-  subject_title_template?: string | null;
-  filter_predicate?: Record<string, unknown>;
-  schedule_config?: { cron?: string; timezone?: string };
 }
 
 interface TriggersTabProps {
