+++ app/frontend/pages/Projects/Workflows/types.ts
@@ -1,0 +1,16 @@
+export interface Trigger {
+  id: number;
+  kind: string;
+  event_type: string;
+  name?: string | null;
+  trigger_mode?: string;
+  cooldown_seconds?: number;
+  enabled?: boolean;
+  column_name?: string;
+  board_column_id?: number;
+  subject_policy?: string;
+  subject_column_id?: number | null;
+  subject_title_template?: string | null;
+  filter_predicate?: Record<string, unknown>;
+  schedule_config?: { cron?: string; timezone?: string };
+}
