+++ app/frontend/shared/resources/assets/AssetsContent.tsx
@@ -30,34 +30,9 @@ import { EmptyState } from 'shared/ui/EmptyState';
 import { PageHeader } from 'shared/ui/PageHeader';
 
 import { AssetPreviewModal } from './AssetPreviewModal';
+import type { Asset, AssetVersion } from './types';
 
-export interface AssetVersion {
-  id: number;
-  version: number;
-  contentType: string | null;
-  fileSize: number | null;
-  source: string | null;
-  fileUrl: string | null;
-  createdAt: string | null;
-}
-
-export interface Asset {
-  id: number;
-  name: string;
-  folder: string | null;
-  tags: string[];
-  public: boolean;
-  scopeType: string;
-  scopeId: number;
-  scopeIndicator: 'company' | 'project';
-  status: string;
-  createdById: number;
-  createdByName: string | null;
-  versionsCount: number;
-  latestVersion: AssetVersion | null;
-  createdAt: string;
-  updatedAt: string;
-}
+export type { Asset, AssetVersion } from './types';
 
 const PRESIGN_URL = '/api/v1/assets/presign';
 const MAX_FILE_SIZE = 1024 * 1024 * 1024;
