+++ app/frontend/shared/resources/mcp-servers/McpServersContent.tsx
@@ -37,42 +37,9 @@ import { McpServerFormModal } from './McpServerFormModal';
 import { driftedServers } from './serverHealth';
 import { ServerHealthIcons } from './ServerHealthIcons';
 import { ToolDriftModal } from './ToolDriftModal';
+import type { McpServer } from './types';
 
-type McpServerKind = 'internal' | 'custom';
-type Transport = 'http' | 'sse' | 'stdio';
-
-export interface McpServer {
-  id: number;
-  name: string;
-  url: string | null;
-  transport: Transport;
-  headers: Record<string, string> | null;
-  description: string | null;
-  kind: McpServerKind;
-  scopeType: string | null;
-  scopeId: number | null;
-  scopeIndicator: string;
-  enabled: boolean;
-  internal: boolean;
-  command: string | null;
-  env: Record<string, string> | null;
-  // OAuth (oauth-unification §4.3/§4.6). oauthStatus is per-current-user and read-only.
-  authType?: 'none' | 'static' | 'oauth';
-  credentialScope?: 'shared' | 'per_user';
-  oauthStatus?: 'pending' | 'active' | 'expiring' | 'error' | null;
-  // Connector provenance and tool-baseline health. Null/false on hand-authored
-  // servers — nobody promised anything about a server someone typed in.
-  connectorName?: string | null;
-  connectorVersion?: string | null;
-  connectorStatus?: 'active' | 'deprecated' | 'deleted' | null;
-  connectorVersionPinned?: boolean;
-  toolBaseline?: boolean;
-  toolDrift?: { added?: string[]; removed?: string[]; changed?: string[]; detected_at?: string } | null;
-  /** Version the catalog now carries, when it differs from the installed one. */
-  connectorUpdateVersion?: string | null;
-  createdAt: string;
-  updatedAt: string;
-}
+export type { McpServer } from './types';
 
 // Maps a server's per-user oauth_status to a connection badge (functional labels,
 // not colour alone, so the state reads without relying on hue).
